+++ R/rksfun-package.R
@@ -6,5 +6,6 @@
 #' @importFrom ggplot2 scale_colour_gradient
 #' @importFrom ggplot2 scale_fill_gradient
 #' @importFrom ggplot2 scale_fill_gradientn
+#' @importFrom grDevices colorRampPalette
 ## usethis namespace: end
 NULL