+++ NEWS.md
@@ -1,2 +1,4 @@
 # rksfun (development version)
 
+ * Add some palettes as ggplot2 ready scales
+ * Add palette logic from the great `NatParksPalettes` package